Moved the ListWidgetHandler from autokey_treewidget to centralwidget. It is used in the logging list view, and therefore does not belong to the tree view.
